Tarefa1

Description

Realização da Tarefa1

Documentation

generateRandoms :: Int -> Int -> [Int]

A Função generateRandoms dada uma semente, retorna uma lista n inteiros gerada aliatoriamente

generateRandoms n seed = let gen = mkStdGen seed -> cria um gerador aleatorio

in take n $ randomRs (0,99) gen -> pega nos primeiros n elementos numa serie infinita de numeros aleatorios entre 0 e 99

nrAleatorio :: Int -> Int

Número aliatório gerado com uma seed

subLista :: Int -> [a] -> [[a]]

Converte uma lista numa lista de lista de tamanho n

convertPiece :: Int -> Piece

Converte um Integer numa Piece | x == 3 = Food Big -> 3 = Food Big | x >= 0 && x = Food Little - 0 n < 70 <= Food Little | otherwise = Wall -> 70 <= 99 <= Wall

printCorridor :: Corridor -> String

Converte um Corridor numa string

converteCorredor :: [Int] -> Corridor

Converte uma lista de Integers num Corridor

converteLabirinto :: [[Int]] -> Maze

Converte uma lista de listas de Integers num Maze

ghostHome :: Int -> Maze

Formato da casa dos Corridores da casa fantasma. Dependendo do tamanho dos corridores do maze, o tamanho da casa dos fantasmas também varia | No caso de comprimento do labirinto ser par, a entrada da casa fantasma vai ter comprimento 2, de maneira a ficar centralizada | No caso de o comprimento do labirinto ser ímpar, a entrada da casa fantasma vai ter comprimento 3, de maneira a ficar centralizada

walls :: Int -> Corridor

A função walls cria um Corridor cheio de Paredes (para as paredes inferior de superiores)

walls2 :: Maze -> Maze

A função walls2 adiciona paredes laterais à volta do Maze

limMaze :: Int -> Int -> Int -> Maze

Junta as paredes superiores e inferiores com o restante do maze | limMaze x y z = (walls x) : (walls2 (createMiddle (x-2) (y-2) z)) ++ [(walls x)] -> Gerámos (x-2) e (y-2) devido à colocação das paredes no fim e início dos Corridores (excepto o primeiro e último)

createMiddle :: Int -> Int -> Int -> Maze

Gera aliatoriamente as Pieces que constituem os restantes Corridores

unir :: Int -> Int -> Int -> [Corridor]

A Função unir une a 3 partes de um labirinto:

A de cima

A do meio (contituida pelos túneis e pela casa fantasma)

A de baixo (obtida através de tirarmos as outras duas de um labirnto)

Quando o y é par retiramos div y 2 - 2 pois a casa fantasma tem de estar ​ X-1 peças de distância da parede superior

Sendo o y ímpar, o labirinto vai ter o mesmo número de corridores para cima e para baixo do corridor do Túnel (nx+5 pois o labirinto vai ter 5 corridores)

O "nx = div y 2 - 2" é numero de corredores da parte de cima que não são "influenciados" pelo labirinto

ghostpart :: Int -> Int -> Int -> Maze

Trata de juntar o que vem antes e depois com a casa dos fantasmas nos seus varios corredores

taken1 :: [[a]] -> Int -> [a]

Função auxiliar. Utilizamos esta função para taken1 [[]] _ = [] Base da recursão. No caso do n > numero de Corridores do maze, então vai dar um Corridor vazio taken1 (x:xs) n | n == 1 = x No caso de queremos o 1º Corridor | n /= 1 = taken1 xs (n-1) eliminamos o 1º Corridor porque não era esse que procuravamos, chamando a função recursivamente

mazetocorridor :: Int -> Int -> Int -> Maze

Função auxiliar. Retira os Corridores que vão ser afetados pela inserção da casa dos fantasmas no maze

replaceInList :: Int -> Piece -> Corridor -> Corridor

Substitui uma lista de indice 0 por uma peça

replaceIncorridorOdd :: Maze -> Int -> Piece -> Maze

Substitui uma peça num Labirinto para meter os tuneis (Impares) | Num Maze cujo Y seja ímpar apenas vamos inserir um túnel (exatamente a metade)

replaceIncorridorEven :: Maze -> Int -> Piece -> Maze

Substitui uma peça num Labirinto para meter os tuneis (Pares)

generateMaze

Arguments

:: Int
largura

-> Int

-> Int
seed

-> Maze
O Maze Final

Gera o maze | O labirinto deve conter no minimo 15 de comprimento e 10 de altura ec hamamos a função unir que nos devolve um Labirinto (Maze) que cumpre os vários requisitos

imprimeLabirinto :: Int -> Int -> Int -> IO ()

Função auxiliar. Permite um melhor visionamento da função generateMaze com o objetivo de detetar potenciais erros

testaGeraLabirinto :: Int -> Int -> Int -> Bool

Dicionário de variáveis (Para efeitos ilustrativos, nesta parte são melhor definidas as variáveis mais utilizadas ao longo da Tarefa1) :

mazeEven -> Labirinto cujo y é ímpar, ou seja, o tunel vai ter 1 de altura

mazeOdd -> Labirinto cujo y é par , ou seja, o tunel vai ter 2 de altura

ghostH -> A casa fantasma (que varia de comprimento se o x é par ou ímpar)

nxh -> Número de peças no inicio do corridor não afetadas pela introdução da casa fantasma (no caso de x par)

nxt -> Número de peças no fim do corridor não afetadas pela introdução da casa fantasma (no caso de x par)

nxho -> Número de peças no inicio do corridor não afetadas pela introdução da casa fantasma (no caso de x ímpar)

nxto -> Número de peças no fim do corridor não afetadas pela introdução da casa fantasma (no caso de x ímpar)

ghostP -> Lista dos vários corredores onde a GhostHome (casa dos fantasmas) está inserida

t2mazeE -> Tira o Corridor 2 do labirinto resultante da função mazetocorridor de um labirinto par

t3mazeE -> Tira o Corridor 3 do labirinto resultante da função mazetocorridor de um labirinto par

t4mazeE -> Tira o Corridor 4 do labirinto resultante da função mazetocorridor de um labirinto par

t5mazeE -> Tira o Corridor 5 do labirinto resultante da função mazetocorridor de um labirinto par

t2maze0 -> Tira o Corridor 2 do labirinto resultante da função mazetocorridor de um labirinto ìmpar

t3maze0 -> Tira o Corridor 3 do labirinto resultante da função mazetocorridor de um labirinto ìmpar

t4maze0 -> Tira o Corridor 4 do labirinto resultante da função mazetocorridor de um labirinto ìmpar

t5maze0 -> Tira o Corridor 5 do labirinto resultante da função mazetocorridor de um labirinto ìmpar

Notas adicionais :

Ao longo deste trabalho utiliza-se a conotoção Maze para os Labirintos e Corridores para os vários corredores de cada labirinto

Testes locais (servem para determinar se a função funciona ou não)

Teste de comprimento e altura. Se der True então o comprimento e a altura estão certas

altura :: Int -> Int -> Int -> Bool

Testa se a altura está correta

comprimento :: Maze -> Bool

Compara o comprimento de cada corridor. Caso algum seja diferente do anterior dá falso

scnd :: Maze -> Maze

Função auxiliar. Utilizada para dar os segundos corredores dos labirintos