Tarefa2

Description

Realização da Tarefa2

Jogada mudar de orientação

getorientation

Arguments

:: Play
jogada que fizemos

-> State
Estado incial dos jogadores

-> Bool
Verdade se a orientação da jogada for diferente da do player a ser movimentado. Falso caso contrário

Compara a orientação do movimento. Condição para que faça a função otherway.

otherway

Arguments

:: Play
Jogada feita

-> State
Estado inicial dos players

-> State
Estado final dos players depois de o player ser jogado

Muda a orientação do player jogado

Jogada mover por um túnel

testasentido

Arguments

:: Maze
Labirinto dado

-> Coords
Coordenadas do player a ser movimentado

-> Orientation
Orientação do player

-> Bool
Verdade se está virado para o lado certo e se está nas coordenadas certas para o lado que vai virar. Falso caso contrário.

Testa se está na orientação do movimento do tunel

dti

Arguments

:: Maze
Labirinto dado

-> Coords
Coordenadas do player a ser movimentado

-> Bool
Verdade se estiver na coordenadas do túnel da esquerda. Falso caso contrário

Testa se o movimento pode ser executado no túnel da esquerda (da primeira coluna)

dtf

Arguments

:: Maze
Labirinto dado

-> Coords
Coordenadas do player a ser movimentado

-> Bool
Verdade se estiver na coordenadas do túnel da direita. Falso caso contrário

Testa se o movimento pode ser executado no túnel da direita (da última coluna).

descobretunel

Arguments

:: Maze
Labirinto dado

-> Coords
Coordenadas do player que se vai movimentar

-> Bool
Verdade se está nas coordenadas certas. Falso caso contrário

Testa se o pacman está nas coordenadas onde possa atravessar o túnel

fantOutroLado

Arguments

:: Maze
labirinto dado

-> Player
Player movimentado

-> [Player]
Lista de players

-> Bool
Verdade se tiver um fantasma do outro lado, falso caso contrário

Testa se existe um fantasma do outro lado do túnel

testaoutrolado

Arguments

:: Maze
Labirinto dado

-> Coords
Coordenas do player movimentado

-> Coords
Coordenadas do fantasma com que podemos embater

-> Bool
Verdade se estiver nas coordenadas do túnel, falso caso contrário

Função auxiliar da anterior. Dada as coordenadas do player que vamos mover, confirma se existe fantasma do outro lado

tunelmove

Arguments

:: Int
ID do player

-> Maze
Labirinto onde os players vão se movimentar

-> Orientation
Orientação do movimento do player

-> [Player]
Lista de players inicial

-> [Player]
Lista de players resultante

Caso o Player se movimente para um túnel

tunel

Arguments

:: Orientation
Orientação do movimento

-> Maze
Labirinto onde se vão movimentar os players

-> Player
Player que vai se movimentar no túnel

-> Player
Player resultante do movimento

Altera o player que vai se vai movimentar pelo tunel

Jogada embate com um fantasma

ghostmove

Arguments

:: Play
Jogada do player

-> State
Estado incial dos jogadores

-> State
Estado resultante do movimento

Movimento de um player caso tenha um fantasma nas coordenadas para onde se vai mover

alpcgh

Arguments

:: Int
ID do player

-> Maze
Labirinto onde se vão movimentar os pacmans e fantasmas

-> [Player]
Lista de players a ser alterada

-> [Player]
Lista de players resultante da alteração

Divide o caso os casos. Caso o pacman esteja no estado Mega vai chamar uma função, caso contrário outra.

megapacman

Arguments

:: Int
ID do player

-> Maze
Labirinto

-> [Player]
Lista de players inicial

-> [Player]
Lista de players alteradas

Testa o estado dos fantasmas do Labirinto (Se estão Dead ou Alive). Dá o resultado dos 2 casos

fantVivoMorto

Arguments

:: Player
Player movimentado

-> [Player]
Lista de players restantes

-> Bool
Verdade se tiver, falso caso contrário

Testa se tem fantasmas vivos na casa para onde vamos movimentar

comparag

Arguments

:: [Player]
Lista de players (neste caso só é usado para fantasmas)

-> Bool
Verdade se os ghosts mods de todos players for Dead. Falso caso contrário

Diz se uma lista é composta por fantasmas ou não

testghostnear

Arguments

:: Play
Jogada a implementar

-> State
Estado dos players

-> Bool
Se tiver um fantasma para onde o player se vai mover dá True. Caso contrário dá Falsp

Testa se há um fantasma para onde o player se vai mover

verifcaghost

Arguments

:: Coords
Coordenadas de um player

-> [Player]
Lista de jogadores analisada

-> Bool
Verdade se houve algum fantasma. Falso caso contrário.

Verifica numas coordenadas se está lá um fantasma ou não

ghostdead :: Maze -> Player -> Player

Altera cada fantasma. O fantasma vai para o meio do maze e o seu estado para para vivo

Jogada comer comida grande

bigmoveLista

Arguments

:: Int
ID do Player

-> Orientation
Orientação do movimentp

-> [Player]
Lista de PLayers que vai ser alterado

-> [Player]
Resultado da alteração

Caso o Player se movimente para um lugar onde está uma peça Big Food

big

Arguments

:: Orientation
Orientação do movimento

-> Player
Player que vai ser alterado

-> Player
Resultado da alteração

Altera o player afetado pelo movimento | O player (pacman) move-se, é lhe adicionado cinco pontos ao score e o seu estado muda para Mega. No caso dos fantasmas, é diminuida a velocidade e o seu estado passa para Dead

metadevelocidade

Arguments

:: PlayerState
PLayerState que vai ser alterado

-> PlayerState
PLayerState resultante

Diminiu a velocidade de um Player para metade

alterafantasmamorto

Arguments

:: Int
ID do player

-> Maze
Labirinto

-> [Player]
Lista de players a ser alterada

-> [Player]
Lista de players resultante

Altera uma lista de fantasmas mortos caso sejam comidos

Jogada comer comida pequena

littlemove

Arguments

:: Int
ID do player

-> Orientation
orientação do movimento

-> [Player]
Lista inicial de players

-> [Player]
Lista de players resultante (Afeta apenas o player movido)

Função utilizada no caso de o pacman se mover para um "lugar" onde está uma comida pequena (Food Little)

little

Arguments

:: Orientation
Orientação do player

-> Player
Player que vai ser alterado

-> Player
Player alterado

Jogada ir contra a parede

Altera o player afetado pelo movimento | O player move-se e é lhe adicionado um ponto ao score

wallmove

Arguments

:: Int
ID do player

-> Orientation
Orientação do player

-> [Player]
Lista de players inicial

-> [Player]
Lista de players resultante

Caso o Player se movimente para um lugar onde está uma peça Wall

wall

Arguments

:: Orientation
Orientação do movimento

-> Player
Player que vai ser alterado

-> Player
Player resultante da alteração

Jogada ir para casa Empty

Altera o Player que vai movimentar para a para

emptymove

Arguments

:: Int
ID do player

-> Orientation
orientação do movimento

-> [Player]
Lista de jogadores presente no labirinto

-> [Player]
Lista de jogadores alteradas (altera apenas o jogador afetado pela jogada)

Função utilizada no caso de o pacman se mover para um "lugar" onde está uma peça vazia (Empty)

empty

Arguments

:: Orientation
Orientação do movimento

-> Player
Player que vai ser alterado

-> Player
Resultado da alteração

Função play | Função a implentar nesta tarefa. Movimenta um player num labirinto

Função que altera individualmente cada player do playstate | Nestes 2 casos fica igual pois se um jogador movimentar-se para um lugar Empty, nada é alterado

play

Arguments

:: Play
Jogada a implementar

-> State
State incial

-> State
Resultado da movimentação a um certo player

Funções auxiliares

quicksort

Arguments

:: [Player]
Lista de players inicial

-> [Player]
Lista de players ordenada

Ordena uma lista de players pelos seu IDs

compplayers

Arguments

:: Player
1º player

-> Player
2º player

-> Bool
Verdade se o ID de um player for maior menor ou igual ao de outro

Compara os IDs de 2 players

junta

Arguments

:: Player
Player inical

-> State
Estado inicial de players

-> State
Estado de final de players com todos os players juntos

Junta um player a um estado (Lista de players)

getpeca

Arguments

:: Maze
Labirinto ao qual vamos buscar a peça

-> Orientation
Orientação do pacman (dependendo da sua orientação, ou seja para onde o pacman se vai mover, a peça vai mudar)

-> Coords
coordenadas do pacman

-> Piece
peça resultante

Função utilizada para descobrir qual é a peça presente na casa para qual o pacman se vai movimentar

verificaID

Arguments

:: Int
ID do player que vai ser movimentado

-> Player
Player que vai ser testado

-> Bool
Verdade se o ID do jogador corresponder ao ID do player que vai ser movimentado). Falso caso contrário

Verifica se o ID do player que vai ser movimentado (apenas do pacman) é correto ou não

move

Arguments

:: Orientation
Orientação do movimento

-> PlayerState
PlayerState que vai ser alterado

-> PlayerState
PLayerState resultado

Muda as coordenadas de um Player (ou seja "movimenta" um Player) de acordo com a orientação do movimento que vamos executar. | Se a orientação do Player for diferente da orientação dada, apenas se vai mudar a orientação do Player, mantendo-se nas mesma coordenadas

movepl

Arguments

:: Orientation
Orientação do movimento

-> Player
Player que vai ser movimentado

-> Player
Player resultante do movimento

Move o pacman

getcoords

Arguments

:: Player
Player dado

-> Coords
Coordenadas do player

Devolve as coordenadas de um player

transEmptymap :: Maze -> Coords -> Maze

Coloca no maze

transEmptycor :: Corridor -> Int -> Corridor

Substitui em cada corredor a peça de indice 0 por Empty

ghst

Arguments

:: Player
Player dado

-> Bool
Verdade se for fantasma. Falso caso contrário

Testa se um player é fantasma ou não

aumentarpontuacao

Arguments

:: Int
Número pelo qual vamos aumentar a pontução

-> PlayerState
Estado do inicial do Player

-> PlayerState
Estado que é resultado do aumento de pontuação

Aumenta a pontução de um Player por um número variável

verificapacman

Arguments

:: Player
Player que vai ser testado

-> Bool
Verdade caso seja pacman. Falso se não for.

Testa se o Player é pacman ou não (neste caso se é pacman ou ghost)

aumentarpontuacao1 :: Int -> PlayerState -> Int -> PlayerState

Multiplica a pontuação de um player por um número inteiro

verifasefantamasmortos

Arguments

:: Player
Pacman que movemos (utilizamos para obter as suas coordenadas)

-> [Player]
Lista de Fantasmas mortos que existem na nas mesmas coordenadas do pacman

-> Bool
Verdade se não existem fantasmas vivos. Falso caso contrário

Verifica se numa lista de players todos estão mortos ou não

restantes

Arguments

:: [Player]
Lista de todos os players existentes

-> Coords
Coordenadas do player que movemos

-> [Player]
Lista de players que não estão nas coordenadas do player que movemos

Devolve a lista de players que não estão nas coordenadas do pacmana que movemos

filtra

Arguments

:: [Player]
Lista de todos players existentes

-> Coords
Coordenadas do player que movemos

-> [Player]
Lista de players que estão nas coordenadas do player que movemos

Filtra os players existentes nas coordenada onde o jogador está.

compara

Arguments

:: [Player]
Lista de players (neste caso só é usado para fantasmas)

-> Bool
Verdade se os ghosts mods de todos players for Dead. Falso caso contrário

Compara os ghostmods

ghostdead1

Arguments

:: Player
Pacman dado

-> Int
Numero de fantasmas morto na casa para onde o pacman se move

-> Player
Pacman com os pontos adicionados

Adiciona (10*y) pontos ao pacman que come os fantasmas. | y é o número de fantasmas mortos que o pacman vai comer (Número de fantasmas mortos que estão naquelas coordenadas)

testamega

Arguments

:: Player
Player que vai ser testado

-> Bool
Verdade se o estado do pacman for mega. Falso casi contrario

Testa se um player está em estado Mega ou não

aumentapnt

Arguments

:: Player
Player dado

-> Int
Número que vai ser somado à pontuação atual do player

-> Player
Player Resultante

Aumenta a pontução de um Player

matapacman

Arguments

:: Player
Player recebido

-> Player
Player

Muda o estado do pacman quando este embate num fantasma (este ultimo mantem-se no mesmo estado)

matapacmanlista

Arguments

:: Int
ID do player

-> [Player]
Lista de players

-> [Player]
Lista de players modificada

Para "matar" o pacman que se moveu contra um fantasma

testafoodbig

Arguments

:: Maze
Labirinto dado

-> Coords
Coordenadas da peça movida

-> Bool
Verdade se está lá uma Food Big falso caso contrario

Testa se tem comida grande no sitio para onde o pacman se moveu

eid :: Int -> [Player] -> Player

Função que encontra o id de um player

maze1 :: Maze

-----------------------------------------------NOTAS------------------------------------------

>NESTE TRABALHO CONSIDEROU-SE QUE OS PACMANS ERAM CRIADOS EM LOCAIS ONDE HAVIAM PEÇAS VAZIAS, OU COMIDA PEQUENA OU COMIDA GRANDE, DAÍ QUE APAGA SEMPRE AS PEÇAS (TIRANDO FANTASMAS) POR ONDE O PACMAN PASSA. AO SER CRIADO NESTE LOCAL O PACMAN NÃO ABSORVE OS PONTOS DO PRÓPRIO LOCAL, APENAS RETIRA A PEÇA.

> NÃO TESTAMOS NEM CRIAMOS UMA FUNÇÃO QUE DIFERENCIE OS IDS, JÁ QUE SUPOMOS QUE SÃO DE ORDEM CRESCENTE E QUE SÃO DIFERENTES

> TIVEMOS VÁRIOS PROBLEMAS COM O CODEBOARD, DAÍ A EXTENSA QUANTIDADE DE TESTES LOCAIS QUE REALIZAMOS

-----------------------------------------------DICIONÁRIO DE VARIÁVEIS------------------------------------------

testaID -> Testa se o id do pacman que vamos movimentar corresponde ao do primeiro pacman da lista de players coords -> Coordenadas do player que vamos movimentar testetunel -> Testa se o player está nas condições certas para atravessar o túnel getorirnt -> Compara a orientação do movimento com o do pacman que vamos mover pl1 -> Player que vamos mover pl2 -> Lista de players retirando o player que vams mover pecaseguinte -> peça que está na coordenadas para onde o pacman se vai mover testafantasma -> testa se existe um fantas para as coordenadas para onde o pacman se vai mover testfant -> Testa se existe um fantasma do outro lado do túnel

Labirinto para o teste 1 (beta)

pcs :: State

Teste Beta 1

mazemd :: Maze

Maze mudar de direção

pcsmd :: State

Teste mudar de direção

mazecd :: Maze

Maze comer comida

pcmcd :: State

Teste comer comida

mazetp :: Maze

Maze tunel par

mazeti :: Maze

Maze tunel impar

pcmtpi :: State

Teste tunel impar

pcmtpf :: State

Teste tunel par

mazecfm :: Maze

Maze teste normal e mega

pcmcfm :: State

Teste Embate Fantasma

pcmcfv :: State

Teste embate fantasma vivo

mazetest :: Maze

Maze testes professores

testet :: State

Teste Fantasma tunel outro lado

pcteste :: State

teste exemplo professores

pcteste1 :: State

teste embate fantasma 1

pcteste2 :: State

teste embate fantasma 2

teste1 :: State

teste embate fantasma 3

teste2 :: State

teste embate fantasma 4

teste3 :: State

teste embate fantasma 5

teste4 :: State

teste embate fantasma 6

teste5 :: State

teste embate fantasma 7

testet1 :: State

teste embate fantasma 8

testefinal :: State

Teste final com todas as situações