Tarefa3

Description

Realização da Tarefa3

Documentation

corridorinstruction :: Corridor -> Instruction

mazeToInst :: Maze -> Instructions

A função mazeToInst converte um Maze numa Instruction , usando a corridorinstructionem todos os corridors do Maze.

indexOf

Arguments

:: Instruction
Elemento para procurar

-> [(Int, Instruction)]
Lista onde se vai procurar esse elemento

-> Int
Dá o indice do elemento ou se nao houver nenhum igual, dá um um numero negativo

A função indexOf dá o indice de um elemento de uma lista.

findRepeats

Arguments

:: [(Int, Instruction)]

-> [(Int, Instruction)]
Pedimos duas Listas iguais para as trabalhar com ambas

-> [(Int, Instruction)]
Dá a Lista final que queremos, mas com o indice atras

A função findRepeats substitui os os elementos iguais por Repeats

cutIndex

Arguments

:: [(Int, Instruction)]
A lista dividida em (indice, Instruction)

-> Instructions
O indice é retirado

A função cutIndex corta o indice que está nas nossas lista, para obtermos a lista pretendida

compactMaze :: Maze -> Instructions

A função compactMaze é a função final que junta tudo.